improved LQ tests

+++ statTools.py
@@ -7,7 +7,7 @@ standard deviation
 """
 
 
-def lower_quartile(num_list:list)->int:
+def lower_quartile(num_list: list)->float:
     num_list.sort()
     return num_list[1]
